fork: adapt to v0.5.5y base + wire patched engine
Post-rebase adaptations for the v0.5.5y (VERSION_CODE 63) upstream tag:

- honestWgStatusId: firestack proxy status ids became Int in v0.5.5y
  (upstream d238191dc changed them from Long); migrate the helper and its
  callers accordingly.
- app/build.gradle: drop a duplicated java toolchain block introduced by the
  rebase auto-merge.
- gradle.properties: firestackRepo=local — resolve the self-built patched
  engine (firestack 310d7bc603, the v0.5.5y pin, + the DoH idle-pool fix
  cherry-picked; celzero/firestack#241) from app/libs/tun2socks.aar via the
  root flatDir repo. One line back to ossrh restores the stock engine.

+++ app/src/full/java/com/celzero/bravedns/util/UIUtils.kt
@@ -160,7 +160,8 @@ object UIUtils {
     // can't be mistaken for "increasing now".
     private const val WG_RX_SAMPLE_STALE_MS = 90_000L
 
-    fun honestWgStatusId(key: String, rawStatusId: Long?, stats: RouterStats?): Long? {
+    // (v0.5.5y: firestack proxy status ids became Int — see upstream d238191dc)
+    fun honestWgStatusId(key: String, rawStatusId: Int?, stats: RouterStats?): Int? {
         if (stats == null) return rawStatusId
         val now = System.currentTimeMillis()
         val prev = wgRxSamples.put(key, WgRxSample(stats.rx, now))
